requirements.txt
python-dotenv==0.21.0
requests==2.28.1
ok-api==1.0.1
